requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
bleach==1.5.0
bs4==0.0.1
builtwith==1.3.3
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
cssselect==1.0.3
decorator==4.1.2
Django==1.8.17
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
html5lib==1.0.1
idna==2.6
image==1.5.20
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
iso8601==0.1.12
Jinja2==2.9.6
js2xml==0.3.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
lxml==3.8.0
m3u8==0.3.5
MarkupSafe==1.0
mechanize==0.3.6
mistune==0.7.4
Naked==0.1.31
nbconvert==5.2.1
nbformat==4.4.0
notebook==5.0.0
numpy==1.14.2
olefile==0.44
pandocfilters==1.4.2
path.py==10.3.1
pathlib2==2.3.0
pickleshare==0.7.4
Pillow==4.2.1
ply==3.11
prompt-toolkit==1.0.15
pyasn1==0.4.2
pycrypto==2.6.1
pyDes==2.0.1
Pygments==2.2.0
pytesseract==0.2.0
python-dateutil==2.6.1
pyv8==1.0
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
scandir==1.5
selenium==3.11.0
shellescape==3.4.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
slimit==0.8.1
testpath==0.3
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
whois==0.7
widgetsnbextension==3.0.2
win-unicode-console==0.5
wincertstore==0.2
